Fix error preventing getting correct answer
The problem was that when finding the rightmost letter, if there are
multiple occurrences of a letter, the previous code would find the index
of the first one from the right. This is incorrect when trying to find
the RIGHTMOST letter, but correct for leftmost letter. I copied and
pasted the leftmost code into the rightmost code, which is why I made
this error. The correct way would be to find the index of the letter
on the rightmost side. This is done with rfind().