I tried...
This problem was so difficult for me. I understand how to solve it. That
is pretty easy to figure out. I just didn't know how to code it. I found
a solution online, which is really elegant. I understand it now too. It
finds the location of the symbols, and puts it in a dictionary with an
empty list. Then it finds tuple location of a box containing the numbers
and the stuff around it. It ands this set of tuples with the symbol
tuple. If there is a remaining tuple, that number is a part number. We
then insert it into the array for the symbol tuple dictionary. Sum
everything. Done. Part 2 came easily because the designers made it
simple.