Part 2 is difficult!
I manually solved this problem by slowly narrowing down the range of
seeds until I can get to a step of 1 with decent computing time. This
took me a bit of time to test inputs and find good enough values for
the final answer. I definitely need to come back to this problem.